Refactored testLib.py
