+++ package.json
@@ -32,5 +32,6 @@
     "test_dfgm": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 test/test_dfgm.py",
     "test_iris": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 test/test_iris.py",
     "test_core_flatsat": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 test/test_flatsat_core_comp.py",
+    "test_hk": "LD_LIBRARY_PATH=./libcsp/build PYTHONPATH=./libcsp/build python3 test/test_full_hk.py"
   }
 }
