+++ test/nim/expected_nim_hk.py
@@ -1,0 +1,29 @@
+'''
+ * Copyright (C) 2022  University of Alberta
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+'''
+'''
+ * @file expected_nim_hk.py
+ * @author Jenish Patel
+ * @date 2022-08-08
+'''
+
+expected_NIM_HK = {
+    # PCB Temperature 00 - Main Board in deg C
+    # PCB Temperature 01 - Main Board in deg C
+    # PCB Temperature 02 - Screen board in deg C
+    # PCB Temperature 03 - Screen board in deg C
+    # PCB Temperature 04 - Camera board in deg C
+    # PCB Temperature 05 - Camera board in deg C
+    # eNIM Pre-Test Backgronud Reading in Lux
+    # eNIM Test Pattern Reading in Lux
+}
