+++ test/solar/expected_solarPanel_hk.py
@@ -1,0 +1,69 @@
+'''
+ * Copyright (C) 2022  University of Alberta
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+'''
+'''
+ * @file expected_solarPanel_hk.py
+ * @author Jenish Patel
+ * @date 2022-08-08
+'''
+
+expected_solarPanel_HK = {
+    'Nadir_Temp1': [0, 70],  # Nadir Temp 1 in deg C
+    'Nadir_Temp_Adc': [0, 70],  # Nadir Temp ADC in deg C
+    'Port_Temp1': [0, 70],  # Port Temp 1-3  in deg C
+    'Port_Temp2': [0, 70],
+    'Port_Temp3': [0, 70],
+    'Port_Temp_Adc': [0, 70],  # Port Temp ADC in deg C
+    'Port_Dep_Temp1': [0, 70],  # Port Dep Temp 1-3 in deg C
+    'Port_Dep_Temp2': [0, 70],
+    'Port_Dep_Temp3': [0, 70],
+    'Port_Dep_Temp_Adc': [0, 70],  # Port Dep Temp ADC in deg C
+    'Star_Temp1': [0, 70],  # Star Temp 1-3 in deg C
+    'Star_Temp2': [0, 70],
+    'Star_Temp3': [0, 70],
+    'Star_Temp_Adc': [0, 70],  # Star Temp ADC in deg C
+    'Star_Dep_Temp1': [0, 70],  # Star Dep Temp 1-3 in deg C
+    'Star_Dep_Temp2': [0, 70],
+    'Star_Dep_Temp3': [0, 70],
+    'Star_Dep_Temp_Adc': [0, 70],  # Star Dep Temp ADC in deg C
+    'Zenith_Temp1': [0, 70],  # Zenith Temp 1-3 in deg C
+    'Zenith_Temp2': [0, 70],
+    'Zenith_Temp3': [0, 70],
+    'Zenith_Temp_Adc': [0, 70],  # Zenith Temp ADC in deg C
+    'Nadir_Pd1': [0, 100],  # Nadir PD 1 in %
+    'Port_Pd1': [0, 100],  # Port PD 1-3 in %
+    'Port_Pd2': [0, 100],
+    'Port_Pd3': [0, 100],
+    'Port_Dep_Pd1': [0, 100],  # Port Dep PD 1-3 in %
+    'Port_Dep_Pd2': [0, 100],
+    'Port_Dep_Pd3': [0, 100],
+    'Star_Pd1': [0, 100],  # Star PD 1-3 in %
+    'Star_Pd2': [0, 100],
+    'Star_Pd3': [0, 100],
+    'Star_Dep_Pd1': [0, 100],  # Star Dep PD 1-3 in %
+    'Star_Dep_Pd2': [0, 100],
+    'Star_Dep_Pd3': [0, 100],
+    'Zenith_Pd1': [0, 100],  # Zenith PD 1-3 in %
+    'Zenith_Pd2': [0, 100],
+    'Zenith_Pd3': [0, 100],
+    'Port_Voltage': [0, 16500],  # Port Voltage in mV
+    'Port_Dep_Voltage': [0, 16500],  # Port Dep Voltage in mV
+    'Star_Voltage': [0, 16500],  # Star Voltage in mV
+    'Star_Dep_Voltage': [0, 16500],  # Star Dep Voltage in mV
+    'Zenith_Voltage': [0, 16500],  # Zenith Voltage in mV
+    'Port_Current': [0, 600],  # Port Current in mA
+    'Port_Dep_Current': [0, 600],  # Port Dep Current in mA
+    'Star_Current': [0, 600],  # Star Current in mA
+    'Star_Dep_Current': [0, 600],  # Star Dep Current in mA
+    'Zenith_Current': [0, 600],  # Zenith Current in mA
+}
