+++ test/yukon/expected_yukon_hk.py
@@ -1,0 +1,27 @@
+'''
+ * Copyright (C) 2022  University of Alberta
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+'''
+'''
+ * @file expected_yukon_hk.py
+ * @author Jenish Patel
+ * @date 2022-08-08
+'''
+
+expected_yukon_HK = {
+    # Temperatures TBD in deg C
+    # Motor Encoder Positions
+    # Payload MCU RAM Available in Bytes
+    # Science Package Data File Size in Bytes
+    # TBD Camera Health
+    # TBD SD Artwork Buffer Positions in "16 Bytes"
+}
